move models into dedicated directory, split in multiple files and make magic to pick them up automatically
